requirements.txt
markdown==3.5.2
pyyaml==6.0.1
Pillow==10.2.0  # For image processing
pygments==2.17.2  # For code highlighting 